+++ actionmailbox/test/unit/mail_ext/addresses_test.rb
@@ -10,15 +10,16 @@ module MailExt
         to: "david@basecamp.com",
         cc: "jason@basecamp.com",
         bcc: "andrea@basecamp.com",
-        x_original_to: "ryan@basecamp.com"
+        x_original_to: "ryan@basecamp.com",
+        x_forwarded_to: "jane@example.com"
     end
 
     test "from address uses address object" do
       assert_equal "example.com", @mail.from_address.domain
     end
 
-    test "recipients include everyone from to, cc, bcc, and x-original-to" do
-      assert_equal %w[ david@basecamp.com jason@basecamp.com andrea@basecamp.com ryan@basecamp.com ], @mail.recipients
+    test "recipients include everyone from to, cc, bcc, x-original-to, and x-forwarded-to" do
+      assert_equal %w[ david@basecamp.com jason@basecamp.com andrea@basecamp.com ryan@basecamp.com jane@example.com ], @mail.recipients
     end
 
     test "recipients addresses use address objects" do
@@ -39,6 +40,10 @@ module MailExt
 
     test "x_original_to addresses use address objects" do
       assert_equal "basecamp.com", @mail.x_original_to_addresses.first.domain
+    end
+
+    test "x_forwarded_to addresses use address objects" do
+      assert_equal "example.com", @mail.x_forwarded_to_addresses.first.domain
     end
   end
 end