+++ actionmailbox/CHANGELOG.md
@@ -1,3 +1,5 @@
+*   Add X-Forwarded-To addresses to recipients.
 
+    *Andrew Stewart*
 
 Please check [7-0-stable](https://github.com/rails/rails/blob/7-0-stable/actionmailbox/CHANGELOG.md) for previous changes.